requirements.txt
# Flask web framework
Flask>=2.0.0
flask-cors>=3.0.10

# Test Generation
pynguin>=0.30.0
openai>=1.0.0

# Git operations
GitPython>=3.1.0

# Utilities
Werkzeug>=2.0.0
python-dotenv>=0.19.0
